sm8650-common: init: Further optimize input boost and up/down rate limit
In my simple test, I found that high frequency does not necessarily mean power consumption.
I tested it for 1 min and the average current was as follows:
...
1248000 -> 142 mA
1344000 -> 142 mA
1459200 -> 138 mA
1574400 -> 140 mA
1689600 -> 137 mA
1804800 -> 140 mA
...
So I finally chose 1689600 as the frequency of input boost and reduced the duration.

And I adjusted the up/down rate limit, so that the frequency of cluster 1-3 would rise quickly
and fall slowly, make cluster 4 a little lazier.